conn_pool: moving APIs to include
