+++ include/envoy/router/router.h
@@ -9,15 +9,18 @@
 #include <string>
 
 #include "envoy/access_log/access_log.h"
+#include "envoy/common/conn_pool.h"
 #include "envoy/common/matchers.h"
 #include "envoy/config/core/v3/base.pb.h"
 #include "envoy/config/route/v3/route_components.pb.h"
 #include "envoy/config/typed_metadata.h"
 #include "envoy/http/codec.h"
 #include "envoy/http/codes.h"
+#include "envoy/http/conn_pool.h"
 #include "envoy/http/hash_policy.h"
 #include "envoy/http/header_map.h"
 #include "envoy/router/internal_redirect.h"
+#include "envoy/tcp/conn_pool.h"
 #include "envoy/tracing/http_tracer.h"
 #include "envoy/type/v3/percent.pb.h"
 #include "envoy/upstream/resource_manager.h"
@@ -32,7 +35,8 @@ namespace Envoy {
 
 namespace Upstream {
 class ClusterManager;
-}
+class LoadBalancerContext;
+} // namespace Upstream
 
 namespace Router {
 
@@ -1090,6 +1094,153 @@ public:
 };
 
 using ConfigConstSharedPtr = std::shared_ptr<const Config>;
+
+class GenericConnectionPoolCallbacks;
+class UpstreamRequest;
+class GenericUpstream;
+
+/**
+ * An API for wrapping either an HTTP or a TCP connection pool.
+ *
+ * The GenericConnPool exists to create a GenericUpstream handle via a call to
+ * newStream resulting in an eventual call to onPoolReady
+ */
+class GenericConnPool {
+public:
+  virtual ~GenericConnPool() = default;
+
+  /**
+   * Called to create a new HTTP stream or TCP connection for "CONNECT streams".
+   *
+   * The implementation of the GenericConnPool will either call
+   * GenericConnectionPoolCallbacks::onPoolReady
+   * when a stream is available or GenericConnectionPoolCallbacks::onPoolFailure
+   * if stream creation fails.
+   *
+   * The caller is responsible for calling cancelAnyPendingRequest() if stream
+   * creation is no longer desired. newStream may only be called once per
+   * GenericConnPool.
+   *
+   * @param callbacks callbacks to communicate stream failure or creation on.
+   */
+  virtual void newStream(GenericConnectionPoolCallbacks* callbacks) PURE;
+  /**
+   * Called to cancel any pending newStream request,
+   */
+  virtual bool cancelAnyPendingRequest() PURE;
+  /**
+   * @return optionally returns the protocol for the connection pool.
+   */
+  virtual absl::optional<Http::Protocol> protocol() const PURE;
+  /**
+   * @return optionally returns the host for the connection pool.
+   */
+  virtual Upstream::HostDescriptionConstSharedPtr host() const PURE;
+};
+
+/**
+ * An API for wrapping callbacks from either an HTTP or a TCP connection pool.
+ *
+ * Just like the connection pool callbacks, the GenericConnectionPoolCallbacks
+ * will either call onPoolReady when a GenericUpstream is ready, or
+ * onPoolFailure if a connection/stream can not be established.
+ */
+class GenericConnectionPoolCallbacks {
+public:
+  virtual ~GenericConnectionPoolCallbacks() = default;
+
+  /**
+   * Called to indicate a failure for GenericConnPool::newStream to establish a stream.
+   *
+   * @param reason supplies the failure reason.
+   * @param transport_failure_reason supplies the details of the transport failure reason.
+   * @param host supplies the description of the host that caused the failure. This may be nullptr
+   *             if no host was involved in the failure (for example overflow).
+   */
+  virtual void onPoolFailure(ConnectionPool::PoolFailureReason reason,
+                             absl::string_view transport_failure_reason,
+                             Upstream::HostDescriptionConstSharedPtr host) PURE;
+  /**
+   * Called when GenericConnPool::newStream has established a new stream.
+   *
+   * @param upstream supplies the generic upstream for the stream.
+   * @param host supplies the description of the host that will carry the request. For logical
+   *             connection pools the description may be different each time this is called.
+   * @param upstream_local_address supplies the local address of the upstream connection.
+   * @param info supplies the stream info object associated with the upstream connection.
+   */
+  virtual void onPoolReady(std::unique_ptr<GenericUpstream>&& upstream,
+                           Upstream::HostDescriptionConstSharedPtr host,
+                           const Network::Address::InstanceConstSharedPtr& upstream_local_address,
+                           const StreamInfo::StreamInfo& info) PURE;
+
+  // TODO(alyssawilk) This exists because the Connection Pool creates the GenericUpstream, and the
+  // GenericUpstream needs a handle back to the upstream request to pass on events, as upstream
+  // data flows in. Do interface clean up in a follow-up PR.
+  virtual UpstreamRequest* upstreamRequest() PURE;
+};
+
+/**
+ * An API for sending information to either a TCP or HTTP upstream.
+ *
+ * It is similar logically to RequestEncoder, only without the getStream interface.
+ */
+class GenericUpstream {
+public:
+  virtual ~GenericUpstream() = default;
+  /**
+   * Encode a data frame.
+   * @param data supplies the data to encode. The data may be moved by the encoder.
+   * @param end_stream supplies whether this is the last data frame.
+   */
+  virtual void encodeData(Buffer::Instance& data, bool end_stream) PURE;
+  /**
+   * Encode metadata.
+   * @param metadata_map_vector is the vector of metadata maps to encode.
+   */
+  virtual void encodeMetadata(const Http::MetadataMapVector& metadata_map_vector) PURE;
+  /**
+   * Encode headers, optionally indicating end of stream.
+   * @param headers supplies the header map to encode.
+   * @param end_stream supplies whether this is a header only request.
+   */
+  virtual void encodeHeaders(const Http::RequestHeaderMap& headers, bool end_stream) PURE;
+  /**
+   * Encode trailers. This implicitly ends the stream.
+   * @param trailers supplies the trailers to encode.
+   */
+  virtual void encodeTrailers(const Http::RequestTrailerMap& trailers) PURE;
+  /**
+   * Enable/disable further data from this stream.
+   */
+  virtual void readDisable(bool disable) PURE;
+  /**
+   * Reset the stream. No events will fire beyond this point.
+   * @param reason supplies the reset reason.
+   */
+  virtual void resetStream() PURE;
+};
+
+using GenericConnPoolPtr = std::unique_ptr<GenericConnPool>;
+
+/*
+ * A factory for creating generic connection pools.
+ */
+class GenericConnPoolFactory : public Envoy::Config::TypedFactory {
+public:
+  ~GenericConnPoolFactory() override = default;
+
+  /*
+   * @param options for creating the transport socket
+   * @return may be null
+   */
+  virtual GenericConnPoolPtr createGenericConnPool(Upstream::ClusterManager& cm, bool is_connect,
+                                                   const RouteEntry& route_entry,
+                                                   Http::Protocol protocol,
+                                                   Upstream::LoadBalancerContext* ctx) const PURE;
+};
+
+using GenericConnPoolFactoryPtr = std::unique_ptr<GenericConnPoolFactory>;
 
 } // namespace Router
 } // namespace Envoy
