+++ test/mocks/router/mocks.h
@@ -519,5 +519,29 @@ public:
   std::shared_ptr<MockScopedConfig> config_;
 };
 
+class MockGenericConnPool : public GenericConnPool {
+  MOCK_METHOD(void, newStream, (GenericConnectionPoolCallbacks * request));
+  MOCK_METHOD(bool, cancelAnyPendingRequest, ());
+  MOCK_METHOD(absl::optional<Http::Protocol>, protocol, (), (const));
+  MOCK_METHOD(bool, initialize,
+              (Upstream::ClusterManager&, const RouteEntry&, Http::Protocol,
+               Upstream::LoadBalancerContext*));
+  MOCK_METHOD(Upstream::HostDescriptionConstSharedPtr, host, (), (const));
+};
+
+class MockGenericConnectionPoolCallbacks : public GenericConnectionPoolCallbacks {
+public:
+  MOCK_METHOD(void, onPoolFailure,
+              (Http::ConnectionPool::PoolFailureReason reason,
+               absl::string_view transport_failure_reason,
+               Upstream::HostDescriptionConstSharedPtr host));
+  MOCK_METHOD(void, onPoolReady,
+              (std::unique_ptr<GenericUpstream> && upstream,
+               Upstream::HostDescriptionConstSharedPtr host,
+               const Network::Address::InstanceConstSharedPtr& upstream_local_address,
+               const StreamInfo::StreamInfo& info));
+  MOCK_METHOD(UpstreamRequest*, upstreamRequest, ());
+};
+
 } // namespace Router
 } // namespace Envoy
